# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: Order Processing API
  version: 1.1.0
  description: Customer/admin orders with immutable base catalog prices and expiring
    region-selected quotes. All responses include a server-generated X-Request-ID.
    Managed rates are not live market data.
servers:
- url: "/"
security:
- BearerAuth: []
paths:
...
  "/api/v1/orders":
    get:
      summary: List own orders (customer) or all orders (admin)
      parameters:
      - "$ref": "#/components/parameters/Limit"
      - "$ref": "#/components/parameters/Cursor"
      - name: status
        in: query
        schema:
          "$ref": "#/components/schemas/Status"
      responses:
        '200':
          description: Newest orders first
          content:
            application/json:
              schema:
                type: object
                required:
                - items
                - next_cursor
                properties:
                  items:
                    type: array
                    items:
                      "$ref": "#/components/schemas/Order"
                  next_cursor:
                    type: string
                    format: uuid
                    nullable: true
        '401':
          "$ref": "#/components/responses/Error"
        '422':
          "$ref": "#/components/responses/Error"
    post:
      summary: Create a base-currency order or consume an owned quote
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
              - type: object
                additionalProperties: false
                required:
                - items
                properties:
                  items:
                    type: array
                    minItems: 1
                    maxItems: 100
                    items:
                      type: object
                      additionalProperties: false
                      required:
                      - product_id
                      - quantity
                      properties:
                        product_id:
                          type: string
                          format: uuid
                        quantity:
                          type: integer
                          format: int64
                          minimum: 1
              - type: object
                additionalProperties: false
                required:
                - quote_id
                properties:
                  quote_id:
                    type: string
                    format: uuid
      responses:
        '201':
          description: Created in PENDING status
          headers:
            Location:
              schema:
                type: string
              description: Order URL
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order"
        '400':
          "$ref": "#/components/responses/Error"
        '401':
          "$ref": "#/components/responses/Error"
        '403':
          "$ref": "#/components/responses/Error"
        '422':
          "$ref": "#/components/responses/Error"
        '200':
          description: Order
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Order"
        '404':
          "$ref": "#/components/responses/Error"
        '409':
          "$ref": "#/components/responses/Error"
      description: Use items alone for a direct base-currency purchase. Use quote_id
        alone for an accepted regional quote. A successful retry of the same quote
        returns the existing order with 200, even after expiry; new orders return
        201.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Cursor:
      name: cursor
      in: query
      description: Last UUID from the previous page; retain other filters
      schema:
        type: string
        format: uuid
  responses:
    Error:
      description: Error (see HTTP status)
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
...
  schemas:
    Error:
      type: object
      required:
      - error
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
...
    Status:
      type: string
      enum:
      - PENDING
      - PROCESSING
      - SHIPPED
      - DELIVERED
      - CANCELLED
...
    Order:
      type: object
      required:
      - id
      - customer_id
      - status
      - currency
      - total_minor
      - items
      - created_at
      - updated_at
      - pricing
      properties:
        id:
          type: string
          format: uuid
        customer_id:
          type: string
          format: uuid
        status:
          "$ref": "#/components/schemas/Status"
        currency:
          type: string
          example: USD
        total_minor:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            "$ref": "#/components/schemas/OrderItem"
        pricing:
          "$ref": "#/components/schemas/Pricing"
...
    Pricing:
      type: object
      required:
      - mode
      - quote_id
      - region
      - mapping_version
      - source_currency
      - source_total_minor
      - base_digits
      - target_digits
      - rate
      - rate_id
      - rate_source
      - rate_valid_from
      - rate_valid_until
      properties:
        mode:
          type: string
          enum:
          - legacy
          - base
          - quote
        quote_id:
          type: string
          nullable: true
          format: uuid
        region:
          type: string
          nullable: true
        mapping_version:
          type: string
          nullable: true
        source_currency:
          type: string
          nullable: true
        source_total_minor:
          type: integer
          nullable: true
          format: int64
        base_digits:
          type: integer
          nullable: true
        target_digits:
          type: integer
          nullable: true
        rate:
          type: string
          nullable: true
          description: Exact decimal target-major/base-major rate
        rate_id:
          type: string
          nullable: true
          format: uuid
        rate_source:
          type: string
          nullable: true
        rate_valid_from:
          type: string
          nullable: true
          format: date-time
        rate_valid_until:
          type: string
          nullable: true
          format: date-time
      description: Immutable provenance. Legacy records have unknown values set to
        null.
    OrderItem:
      type: object
      required:
      - product_id
      - sku
      - name
      - quantity
      - unit_price_minor
      - line_total_minor
      properties:
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        quantity:
          type: integer
          format: int64
        unit_price_minor:
          type: integer
          format: int64
        line_total_minor:
          type: integer
          format: int64
        source_unit_price_minor:
          type: integer
          nullable: true
          format: int64
        source_line_total_minor:
          type: integer
          nullable: true
          format: int64
...